refactor : server > client로 수정
